refactor: extract error handlers into separate handlers.py module
Move error handler functions and related utilities from __init__.py
into a dedicated handlers.py submodule for better organization.

- Create handlers.py with all error handler functions
- Simplify __init__.py to re-export public API
- Maintain backward compatibility with existing imports

This improves code organization by separating handler implementation
from the module's public interface.
